+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Improve how inherited members are detected.
  - Improve error message when module live-reload fails.
  - Smaller favicon, improved CSS minification
+ - Improve error message if module is not found.
 
 # 2021-01-26 pdoc 3.0.1